requirements.txt
jsonpath==0.82
Pillow==8.4.0
opencv-python==3.4.4.19
python-lorem==1.1.2
google-cloud-vision==2.6.3
pytesseract==0.2.7

